src/API/requirements.txt
connexion[swagger-ui] == 2.6.0
python_dateutil >= 2.6.0
setuptools == 21.0.0
psycopg[binary]
simplejson
flask-cors